+++ CHANGELOG.md
@@ -75,6 +75,7 @@ This is the rolling changelog for TShock for Terraria. Use past tense when addin
 * Moved `Utils.Ban()` to `TSPlayer.Ban()`. (@hakusaro)
 * Moved `Utils.SendMultipleMatchError()` to `TSPlayer.SendMultipleMatchError`. (@hakusaro)
 * Removed `Utils.GetPlayers()`. Iterate over the TSPlayers on the server and make your own list.
+* Removed `Utils.HasBanExpired()` and replaced with `Bans.RemoveBanIfExpired()`. (@hakusaro)
 
 ## TShock 4.3.25
 * Fixed a critical exploit in the Terraria protocol that could cause massive unpreventable world corruption as well as a number of other problems. Thanks to @bartico6 for reporting. Fixed by the efforts of @QuiCM, @hakusaro, and tips in the right directioon from @bartico6.